+++ crates/maple_derive/src/impls/clap_plugin.rs
@@ -8,6 +8,7 @@ use proc_macro::{self, TokenStream};
 use proc_macro2::Span;
 use quote::quote;
 use syn::{DeriveInput, Error, Expr, Ident, LitStr};
+use types::PLUGIN_ACTION_SEPARATOR;
 
 static PLUGINS: Lazy<Mutex<HashSet<String>>> = Lazy::new(|| Mutex::new(HashSet::new()));
 
@@ -16,9 +17,6 @@ struct Plugin {
     id: LitStr,
     actions: Option<Expr>,
 }
-
-/// Separator for non-system plugin actions, `git.reload`.
-const SEP: char = '.';
 
 pub fn clap_plugin_derive_impl(input: &DeriveInput) -> TokenStream {
     let mut maybe_plugin_id = None;
@@ -141,7 +139,7 @@ pub fn clap_plugin_derive_impl(input: &DeriveInput) -> TokenStream {
         let namespaced_action = if plugin_id == "system" {
             action.clone()
         } else {
-            format!("{plugin_id}{SEP}{action}")
+            format!("{plugin_id}{PLUGIN_ACTION_SEPARATOR}{action}")
         };
 
         if is_callable {
@@ -174,7 +172,7 @@ pub fn clap_plugin_derive_impl(input: &DeriveInput) -> TokenStream {
             let method = if plugin_id == "system" {
                 arg.to_string()
             } else {
-                format!("{plugin_id}{SEP}{arg}")
+                format!("{plugin_id}{PLUGIN_ACTION_SEPARATOR}{arg}")
             };
             let pascal_name = if let Some(name) = arg.strip_prefix("__") {
                 format!("__{}", to_pascal_case(name))